indexer/requirements.txt
serena-agent @ git+https://github.com/oraios/serena
